Add Pages, Backend (Google-side), Validation etc.
Who thought it was a good idea to make a guy do both back-end and front-end AT THE SAME TIME and then make a Google Cloud API for Google Sheets. AT THE SAME TIME?? Holy hell. One-man army.

Just a note, everytime a form is submitted from the Speed Round page it activates a google trigger script which updates the sheet which automatically updates the points according to spec. Additionally, the leaderboard (in coming time) will update manually from a public .csv file of the sheet. I should get a medal.